+++ src/objc/objc_object.rs
@@ -1,6 +1,15 @@
-use super::Class;
+use super::{sys, Class, Ivar, ObjectType, Sel, BOOL};
 use crate::core::Arc;
-use std::{cell::UnsafeCell, fmt, marker::PhantomData, panic::RefUnwindSafe, ptr::NonNull};
+use std::{
+    alloc,
+    cell::UnsafeCell,
+    ffi::{c_void, CStr},
+    fmt,
+    marker::PhantomData,
+    mem::ManuallyDrop,
+    panic::RefUnwindSafe,
+    ptr::NonNull,
+};
 
 /// An automatically-reference-counted pointer to a type-erased Objective-C
 /// object.
@@ -88,5 +97,52 @@ impl fmt::Debug for ObjCObject<'_> {
     #[inline]
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         (self as *const Self).fmt(f)
+    }
+}
+
+impl ObjCObject<'_> {
+    /// Returns `true` if this class implements or inherits a method that can
+    /// respond to a specified message.
+    ///
+    /// See [documentation](https://developer.apple.com/documentation/objectivec/1418956-nsobject/1418583-respondstoselector).
+    #[inline]
+    #[doc(alias = "respondsToSelector")]
+    pub fn responds_to_selector(&self, selector: Sel) -> bool {
+        unsafe { _msg_send_any_cached![self, respondsToSelector: selector => BOOL] }.into()
+    }
+
+    /// Changes the value of an instance variable of a class instance and
+    /// returns its [`Ivar`].
+    ///
+    /// See [documentation](https://developer.apple.com/documentation/objectivec/1441498-object_setinstancevariable?language=objc).
+    ///
+    /// # Safety
+    ///
+    /// `value` must be the expected memory layout and representation for the
+    /// object's instance variable. This is checked in debug builds.
+    #[inline]
+    #[doc(alias = "object_setInstanceVariable")]
+    pub unsafe fn set_ivar<'a, T>(&'a mut self, name: &CStr, value: T) -> &'a Ivar {
+        let mut value = ManuallyDrop::new(value);
+        let value = &mut *value as *mut T as *mut c_void;
+
+        // Ensure in debug builds that this is kosher.
+        if cfg!(debug_assertions) {
+            let class = self.class();
+            let ty = std::any::type_name::<T>();
+
+            let ivar = class.get_ivar(name).unwrap_or_else(|| {
+                let class = class.name();
+                panic!("No instance variable {name:?} on {class:?}");
+            });
+
+            assert_eq!(
+                ivar.type_encoding().layout(),
+                alloc::Layout::new::<T>(),
+                "Incorrect layout for writing `{ty}` to {name:?} on {class:?}",
+            );
+        }
+
+        &*sys::object_setInstanceVariable(self, name.as_ptr(), value)
     }
 }